test schemas in a separate file
